+++ docs/source/conf.py
@@ -25,7 +25,7 @@ copyright = "2018-2019, GreyNoise Intelligence"
 author = "GreyNoise Intelligence"
 
 # The full version, including alpha/beta/rc tags
-release = "0.5.0"
+release = "release = "0.5.0""
 
 
 # -- General configuration ---------------------------------------------------